# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Synthora API",
    "description": "Multi-model AI orchestration — answers that surpass any single model",
    "version": "2.8.8"
  },
  "paths": {
...
    "/api/history": {
      "get": {
        "summary": "History",
        "description": "Return paginated query history for the current user.",
        "operationId": "history_api_history_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/profile/cognitive-consent": {
      "post": {
        "summary": "Toggle Cognitive Consent",
        "description": "Toggle cognitive tracking consent (opt-in/opt-out).",
        "operationId": "toggle_cognitive_consent_api_profile_cognitive_consent_post",
        "parameters": [
          {
            "name": "consent",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Consent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CognitiveConsentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/profile/language": {
      "get": {
        "summary": "Get Language",
        "description": "Return current user's preferred language.",
        "operationId": "get_language_api_profile_language_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "put": {
        "summary": "Set Language",
        "description": "Persist current user's preferred language.",
        "operationId": "set_language_api_profile_language_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetLanguageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/profile/delete-cognitive": {
      "post": {
        "summary": "Delete Cognitive Data",
        "description": "Delete all cognitive tracking data for the current user (原则#22 用户主权).\n\nPreserves non-cognitive fields (preferences, mode history).\nRequires explicit confirmation via query param ?confirm=true.",
        "operationId": "delete_cognitive_data_api_profile_delete_cognitive_post",
        "parameters": [
          {
            "name": "confirm",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Confirm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteCognitiveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/improvement-plans/{plan_id}/activate": {
      "post": {
        "summary": "Activate Plan",
        "description": "Accept and activate a proposed improvement plan.",
        "operationId": "activate_plan_api_improvement_plans__plan_id__activate_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImprovementPlanActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/improvement-plans/{plan_id}/abandon": {
      "post": {
        "summary": "Abandon Plan",
        "description": "Abandon an improvement plan.",
        "operationId": "abandon_plan_api_improvement_plans__plan_id__abandon_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImprovementPlanActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/improvement-plans/{plan_id}/engaged": {
      "post": {
        "summary": "Record Challenge Engagement",
        "description": "Record that user engaged with a micro-challenge (responded to it).",
        "operationId": "record_challenge_engagement_api_improvement_plans__plan_id__engaged_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImprovementPlanActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CognitiveConsentResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "cognitive_tracking_consent": {
            "type": "boolean",
            "title": "Cognitive Tracking Consent"
          }
        },
        "type": "object",
        "required": [
          "status",
          "cognitive_tracking_consent"
        ],
        "title": "CognitiveConsentResponse"
      },
...
      "DeleteCognitiveResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "remaining_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Remaining Fields"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message",
          "remaining_fields"
        ],
        "title": "DeleteCognitiveResponse"
      },
      "DivergencePosition": {
        "properties": {
          "stance": {
            "type": "string",
            "title": "Stance"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "models": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Models"
          }
        },
        "type": "object",
        "required": [
          "stance",
          "summary",
          "models"
        ],
        "title": "DivergencePosition"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoryItem": {
        "properties": {
          "query_id": {
            "type": "string",
            "title": "Query Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "final_answer": {
            "type": "string",
            "title": "Final Answer"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "contributor_count": {
            "type": "integer",
            "title": "Contributor Count"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          },
          "estimated_cost_usd": {
            "type": "number",
            "title": "Estimated Cost Usd"
          },
          "user_marked_usable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Marked Usable"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "quality_gate": {
            "type": "string",
            "title": "Quality Gate",
            "default": ""
          },
          "best_single_answer": {
            "type": "string",
            "title": "Best Single Answer",
            "default": ""
          },
          "has_divergence": {
            "type": "boolean",
            "title": "Has Divergence",
            "default": false
          },
          "divergence_summary": {
            "type": "string",
            "title": "Divergence Summary",
            "default": ""
          },
          "key_insights": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Insights"
          },
          "divergence_points": {
            "items": {
              "$ref": "#/components/schemas/SocraticRevealDivergencePoint"
            },
            "type": "array",
            "title": "Divergence Points"
          }
        },
        "type": "object",
        "required": [
          "query_id",
          "question",
          "mode",
          "final_answer",
          "confidence",
          "contributor_count",
          "latency_ms",
          "estimated_cost_usd",
          "created_at",
          "key_insights",
          "divergence_points"
        ],
        "title": "HistoryItem"
      },
      "HistoryResponse": {
        "properties": {
          "history": {
            "items": {
              "$ref": "#/components/schemas/HistoryItem"
            },
            "type": "array",
            "title": "History"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "history",
          "total"
        ],
        "title": "HistoryResponse"
      },
      "ImprovementPlanActionResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImprovementPlanRecord"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ImprovementPlanActionResponse"
      },
      "ImprovementPlanRecord": {
        "properties": {
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "current_level": {
            "type": "integer",
            "title": "Current Level"
          },
          "target_level": {
            "type": "integer",
            "title": "Target Level"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "difficulty": {
            "type": "integer",
            "title": "Difficulty"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "last_challenge_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Challenge Date"
          },
          "challenges_delivered": {
            "type": "integer",
            "title": "Challenges Delivered",
            "default": 0
          },
          "challenges_engaged": {
            "type": "integer",
            "title": "Challenges Engaged",
            "default": 0
          },
          "milestones": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Milestones"
          }
        },
        "type": "object",
        "required": [
          "plan_id",
          "topic",
          "current_level",
          "target_level",
          "status",
          "difficulty"
        ],
        "title": "ImprovementPlanRecord"
      },
...
      "SetLanguageRequest": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language"
          }
        },
        "type": "object",
        "required": [
          "language"
        ],
        "title": "SetLanguageRequest"
      },
...
      "SocraticRevealDivergencePoint": {
        "properties": {
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "positions": {
            "items": {
              "$ref": "#/components/schemas/DivergencePosition"
            },
            "type": "array",
            "title": "Positions"
          },
          "consensus_ratio": {
            "type": "number",
            "title": "Consensus Ratio"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          }
        },
        "type": "object",
        "required": [
          "topic",
          "description",
          "positions",
          "consensus_ratio",
          "difficulty"
        ],
        "title": "SocraticRevealDivergencePoint"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}